Item11512: Add MD5 signature function, other fixes.
Added code to detect that the text does not have any <html> tag.  If
not, wrap the text in <html> </html> tags,  run the parser and then
strip the tags.  This seems to fix most of the Ajax issues -
JQueryAjaxHelper, the JQueryTabpane topic,  and some others.

Also add a MD5 function to match inline code against a list of "safe"
signatures.   This will let the plugin work on an older foswiki that has
not remove all of the inline javascript. And it's much less error prone
than crafting Regular Expressions.

git-svn-id: http://svn.foswiki.org/trunk/SafeWikiPlugin@13936 0b4bb1d4-4e5a-0410-9cc4-b2b747904278