Item12327: document newly/added preferred way to check whether SWP can sign stuff
It's not sufficient to check {SafeWikiPlugin}{Enabled} in LSC if a caller
wants to sign code via SWP; after all, an old version of SWP might be
installed. So, instead, document the preferred alternative of checking whether
the 'SafeWikiSignable' context is active (that context got ninja-added in the
previous commit).


git-svn-id: http://svn.foswiki.org/trunk/SafeWikiPlugin@16451 0b4bb1d4-4e5a-0410-9cc4-b2b747904278